+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 
 ## Improvements
 * Change `np.NAN` to `np.nan` to support numpy 2.0 [PR #1245](https://github.com/catalystneuro/neuroconv/pull/1245)
-* Improve error display on scan image interfaces [PR #1246(https://github.com/catalystneuro/neuroconv/pull/1246)
+* Improve error display on scan image interfaces [PR #1246](https://github.com/catalystneuro/neuroconv/pull/1246)
 
 # v0.7.1 (March 5, 2025)